add DeletecUser

+++ src/Modules/CreateInvoiceSystem.Modules.Users.Domain/Application/Commands/DeleteUserCommand.cs
@@ -14,7 +14,10 @@ public class DeleteUserCommand : CommandBase<User, UserDto, IUserRepository>
             throw new ArgumentNullException(nameof(_userRepository));
 
         var userEntity = await _userRepository.GetUserByIdAsync(Parametr.UserId, cancellationToken) ?? throw new InvalidOperationException($"User with ID {Parametr.UserId} not found.");
-        
+
+        if (userEntity.Invoices.Any() || userEntity.Clients.Any() || userEntity.Products.Any())
+            throw new InvalidOperationException($"Cannot delete User with ID {Parametr.UserId} because it has associated Invoices.");
+
         await _userRepository.RemoveAsync(userEntity, cancellationToken);
         if (userEntity.Address is not null)
             await _userRepository.RemoveAddress(userEntity.Address, cancellationToken);
